docs: added the style.txt doc which explains how to style commit messages.
The text explains that a commit should have at least a title that succinctly
explains what the commit is doing. Examples include: feature, fix, docs,
style, refactor, test and chore.  Feature is for adding a new feature, fix
is for fixing a bug, docs is for changes to documentation, style is for:
formatting, missing semi colons, but no code change.  Refactor is for
refactoring production code, test is for adding tests refactoring test; no
production change to code and chore is for updating build tasks, package
manager configs, etc. No prod change.  Each line should be < 72 chars.
Last but not least, focus on WHY the change was made, not how.  The code
explains how the change was made.

The space between the title body and footer are important.  Bullets can also be done like this:
* this
* or this
* or this